+++ bpdm-cleaning-service-dummy/src/test/kotlin/org/eclipse/tractusx/bpdm/cleaning/service/CleaningServiceApiCallsTest.kt
@@ -102,6 +102,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
             .copyWithBpnReferences(BpnReference.empty)
             .copyWithLegalAddress(PostalAddress.empty)
             .copyWithSiteMainAddress(PostalAddress.empty)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -130,6 +135,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
         val mockedBusinessPartner = businessPartnerFactory.createFullBusinessPartner("test")
             .copyWithLegalAddress(PostalAddress.empty)
             .copyWithSiteMainAddress(PostalAddress.empty)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -158,6 +168,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
             .copyWithBpnReferences(BpnReference.empty)
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(site = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -185,6 +200,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
         val mockedBusinessPartner = businessPartnerFactory.createFullBusinessPartner("test")
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(site = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -212,6 +232,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
             .copyWithBpnReferences(BpnReference.empty)
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList()
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -239,6 +264,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
         val mockedBusinessPartner = businessPartnerFactory.createFullBusinessPartner("test")
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -266,6 +296,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
             .copyWithBpnReferences(BpnReference.empty)
             .copyWithSiteMainAddress(null)
             .copy(additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -292,6 +327,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
         val mockedBusinessPartner = businessPartnerFactory.createFullBusinessPartner("test")
             .copyWithSiteMainAddress(null)
             .copy(additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -318,6 +358,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
             .copyWithBpnReferences(BpnReference.empty)
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(site = null, additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)
@@ -345,6 +390,11 @@ class CleaningServiceApiCallsTest @Autowired constructor(
         val mockedBusinessPartner = businessPartnerFactory.createFullBusinessPartner("test")
             .copyWithLegalAddress(PostalAddress.empty)
             .copy(site = null, additionalAddress = null)
+            .copy(
+                uncategorized = businessPartnerFactory.createFullBusinessPartner("test").uncategorized.copy(
+                    identifiers = emptyList() // Assign empty list to uncategorized identifiers
+                )
+            )
 
         val resolveMapping = mockOrchestratorResolveApi()
         mockOrchestratorReserveApi(mockedBusinessPartner)